+++ .github/workflows/pylint.yml
@@ -24,4 +24,4 @@ jobs:
         pip install prt_phasecurve
     - name: Analysing the code with pylint
       run: |
-        pylint $(git ls-files '*.py') --fail-under=9 --ignore-paths=doc,gcmt/tests
+        pylint $(git ls-files '*.py') --fail-under=9 --ignore-paths=doc,gcmt/tests --disable=R0913,R0914,C0415
