+++ gcmt/__init__.py
@@ -1,5 +1,5 @@
 """ init function """
-from .GCMtools import GCMT
+from .gcmtools import GCMT
 from .utils import gcm_plotting
 
 __author__ = 'Aaron David Schneider, Sven Kiefer, Robin Baeyens and Thomas Konings'
